finishing JWT in all, problem at tests

+++ src/main/java/com/yellow/ordermanageryellow/Service/ProductService.java
@@ -1,12 +1,17 @@
 package com.yellow.ordermanageryellow.service;
 
 
+import com.yellow.ordermanageryellow.Service.ProductMapper;
 import com.yellow.ordermanageryellow.dao.ProductRepository;
 import com.yellow.ordermanageryellow.Dto.ProductDTO;
 import com.yellow.ordermanageryellow.Dto.ProductNameDTO;
 import com.yellow.ordermanageryellow.Exception.ObjectAllReadyExists;
-import com.yellow.ordermanageryellow.model.AuditData;
-import com.yellow.ordermanageryellow.model.Product;
+import com.yellow.ordermanageryellow.dao.RolesRepository;
+import com.yellow.ordermanageryellow.exceptions.NoPermissionException;
+import com.yellow.ordermanageryellow.exceptions.ObjectAlreadyExistException;
+import com.yellow.ordermanageryellow.model.*;
+import com.yellow.ordermanageryellow.security.EncryptedData;
+import com.yellow.ordermanageryellow.security.JwtToken;
 import lombok.SneakyThrows;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -17,20 +22,29 @@ import java.util.*;
 @Service
 public class ProductService {
     @Autowired
+    private RolesRepository rolesRepository;
+    @Autowired
+    private JwtToken jwtToken;
+    @Autowired
     private ProductRepository productRepository;
-
-
     public List<Product> getProducts() {
         return productRepository.findAll();
     }
 
     @SneakyThrows
-    public Product addProduct(Product product) throws ObjectAllReadyExists {
-        if (productRepository.existsByName(product.getName()) == true)
-            throw new ObjectAllReadyExists("Object Is All Ready Exists");
-        product.setAuditData(new AuditData(LocalDateTime.now(), null));
-        Product newProduct = productRepository.insert(product);
-        return newProduct;
+    public Product addProduct(Product product,String token) throws ObjectAllReadyExists {
+        String role= this.jwtToken.decryptToken(token, EncryptedData.ROLE);
+        String company= this.jwtToken.decryptToken(token, EncryptedData.COMPANY);
+        Roles wholeRole = rolesRepository.findById(role).orElse(null);
+        if(!wholeRole.getName().equals(RoleNames.ADMIN))
+            throw new NoPermissionException("You do not have permission to delete product category");
+        if (this.productRepository.existsByName(product.getName()))
+            throw new ObjectAlreadyExistException("category name already exist");
+        Company companyOfUser=new Company();
+        companyOfUser.setId(company);
+        product.setCompanyId(companyOfUser);
+        product.setAuditData(new AuditData(LocalDateTime.now()));
+        return this.productRepository.save(product);
 
     }
 
@@ -41,21 +55,34 @@ public class ProductService {
     }
 
     @SneakyThrows
-    public Product editProduct(Product product) {
-        Optional<Product> productOptional = this.productRepository.findById(product.getId());
-
-        if (productOptional.isEmpty())
+    public Product editProduct(Product product, String token) throws ObjectAllReadyExists {
+        String role= this.jwtToken.decryptToken(token, EncryptedData.ROLE);
+        String company= this.jwtToken.decryptToken(token, EncryptedData.COMPANY);
+        Product productOptional = this.productRepository.findById(product.getId()).orElse(null);
+        if (productOptional==null)
             throw new NoSuchElementException("product doesn't exist");
-        Product updatedProduct = productOptional.orElseThrow(() -> new Exception("company not found"));
-        if (!updatedProduct.getName().equals(product.getName()) && productRepository.existsByName(product.getName()))
+        String companyOfProduct = productOptional.getCompanyId().getId();
+        Roles wholeRole = rolesRepository.findById(role).orElse(null);
+        if(!wholeRole.getName().equals(RoleNames.ADMIN)|| !company.equals(companyOfProduct))
+            throw new NoPermissionException("You do not have permission to update product");
+        if (!productOptional.getName().equals(product.getName()) && productRepository.existsByName(product.getName()))
             throw new ObjectAllReadyExists("You need a unique name for product");
-        updatedProduct.getAuditData().setUpdateDate(LocalDateTime.now());
-        updatedProduct = productRepository.save(updatedProduct);
-        return updatedProduct;
+        product.getAuditData().setUpdateDate(LocalDateTime.now());
+       return productRepository.save(product);
     }
+    public void deleteProduct(String id, String token) {
 
-
-    public void deleteProduct(String id) {
+        String role= this.jwtToken.decryptToken(token, EncryptedData.ROLE);
+        String company= this.jwtToken.decryptToken(token, EncryptedData.COMPANY);
+        Product ProductFromDb = this.productRepository.findById(id).orElse(null);
+        if (ProductFromDb == null) {
+            throw new NoSuchElementException("category is not found");
+        }
+        String companyOfCategory = "7";
+                //ProductFromDb.getCompanyId().getId();
+        Roles wholeRole = rolesRepository.findById(role).orElse(null);
+        if( !wholeRole.getName().equals(RoleNames.ADMIN)|| !company.equals(companyOfCategory))
+            throw new NoPermissionException("You do not have permission to delete product category");
         this.productRepository.deleteById(id);
     }
